Add more bugfixes to 5, safekeeping intermediate solution
Not fully resolved due to messages without newline.
However, this likely requires switching from the next line iterators to
finding it manually to filter out when it's not present, as when the
socket is closed there is a flush, and the iterator does not seem to
distinguish between getting a newline and getting a flush.